hien 1

+++ database/migrations/2023_05_15_162821_create_order_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
                 $table->string('product_name');
                 $table->float('product_price');
                 $table->integer('product_quantity');
-                $table->integer('product_feeship');
+                $table->float('product_feeship');
 
                 $table->foreign('order_id')
                 ->references('order_id')
